Move KubeApiClients to its own package
