+++ titus-server-master/src/main/java/com/netflix/titus/master/mesos/kubeapiserver/DefaultKubeJobManagementReconciler.java
@@ -51,7 +51,7 @@ import com.netflix.titus.master.MetricConstants;
 import com.netflix.titus.master.mesos.ContainerEvent;
 import com.netflix.titus.master.mesos.MesosConfiguration;
 import com.netflix.titus.master.mesos.V3ContainerEvent;
-import com.netflix.titus.master.mesos.kubeapiserver.direct.KubeApiFacade;
+import com.netflix.titus.master.mesos.kubeapiserver.client.KubeApiFacade;
 import com.netflix.titus.master.mesos.kubeapiserver.direct.model.PodEvent;
 import com.netflix.titus.master.mesos.kubeapiserver.direct.model.PodNotFoundEvent;
 import io.kubernetes.client.openapi.models.V1Node;
