+++ titus-server-master/src/main/java/com/netflix/titus/master/mesos/kubeapiserver/client/OkHttpMetricsInterceptor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Netflix, Inc.
+ * Copyright 2020 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.master.mesos.kubeapiserver;
+package com.netflix.titus.master.mesos.kubeapiserver.client;
 
 import java.io.IOException;
 import java.time.Duration;
